update the codebase structure
