+++ anchor/programs/gotsol/src/instructions/mod.rs
@@ -1,0 +1,18 @@
+pub mod create_merchant;
+pub mod withdraw_spl;
+pub mod withdraw_sol;
+pub mod refund_spl;
+pub mod refund_sol;
+pub mod set_merchant_status;
+pub mod close_merchant;
+pub mod close_refund;
+
+pub use create_merchant::*;
+pub use withdraw_spl::*;
+pub use withdraw_sol::*;
+pub use refund_spl::*;
+pub use refund_sol::*;
+pub use set_merchant_status::*;
+pub use close_merchant::*;
+pub use close_refund::*;
+
