+++ anchor/programs/gotsol/src/state/mod.rs
@@ -1,0 +1,5 @@
+pub mod merchant;
+pub mod refund_record;
+
+pub use merchant::*;
+pub use refund_record::*;
